Slight changes in path and implementation and change db path to use LOCALAPPDATA instead of roaming

+++ RelayX/core/delete_account.py
@@ -3,7 +3,8 @@ from plyer import notification
 
 
 from RelayX.core.tor_bootstrap import stop_tor
-from RelayX.utils.config import HSDIR, db_path as db_file
+from RelayX.utils.config import db_filepath as db_file
+from utilities.network.network_service import HSDIR
 
 
 async def shutdown_backend(delay : float = 0.5):
@@ -13,8 +14,9 @@ async def shutdown_backend(delay : float = 0.5):
 async def perform_account_deletion():
     try:
         stop_tor()
-        with open(db_file, "w") as f:
-            f.write("")
+        if os.path.exists(db_file):
+            with open(db_file, "w") as f:
+                f.write("")
         data_dir = os.path.join(HSDIR, "..")
         shutil.rmtree(data_dir)
         asyncio.create_task(shutdown_backend())
